created app extender for extending core package settings

+++ src/Strukt/Framework/Console/Command/CacheClear.php
@@ -6,14 +6,32 @@ use Strukt\Console\Input;
 use Strukt\Console\Output;
 
 /**
-* cache:clear     Clear cache
+* cache:clear  Cache clear
+* 
+* Usage:
+*	
+*      cache:clear [<facet>]
+*
+* Arguments:
+*
+*      facet     optional: (app, rtr, ctr, validation, etc..)
 */
 class CacheClear extends \Strukt\Console\Command{
 
 	public function execute(Input $in, Output $out){
 
-		fs()->rmdir(".cache");
+		$facet = $in->get("facet");
+		if(is_null($facet))
+			fs()->rmdir(".cache");
 
-		$out->add("Cache cleared.\n");
+		if(notnull($facet)){
+
+			if(negate(fs(".cache")->isFile(sprintf("%s.json", $facet))))
+				raise(sprintf("File [.cache/%s.json] does not exists!", $facet));
+			
+			fs(".cache")->rm(sprintf("%s.json", strtolower($facet)));
+		}
+
+		$out->add(sprintf("Cache cleared%s.\n", notnull($facet)?sprintf(" [%s.json]", $facet):""));
 	}
 }
